+++ packages/terra-breakpoints/CHANGELOG.md
@@ -3,6 +3,8 @@ ChangeLog
 
 Unreleased
 ----------
+### Fixed
+* Handle the `orientationchange` event on mobile device rotation since the `resize` event is not triggered.
 
 2.26.0 - (April 28, 2020)
 ------------------